requirements.txt
python-doctr[viz]
